feat: Initial test suite, based on the Landmarks extension
This is a basic version of the test suite from Landmarks that has been designed to work with other tools, such as a11y-outline/aria-api. The fixtures and expectations come from the Landmarks browser extension, but the code was written from scratch using TDD.

There are lots more features to come, though this initial version is used as a drop-in replacement for the Landmarks extension test suite.

For more information, consult https://github.com/matatk/landmarks/issues/303