feat(Metadata): Include basic metadata for each test
This is just a human-friendly test name, though other fields could be specified (a "links" field is in-progress). The example programs have been updated to use the new format. A schema and Ajv are are used to validate the expectations against the schema.

Package scripts are added to run the validation (part of testing) and run each of the example scripts.