feat: Build combined fixtures and expectation files
The idea behind this is to allow the tests not only to be run in a browser, but orchestrated from within the browser.

Currently the tests that rely on roles on the body ("application-alone-on-body-is-ignored" and "landmark-role-on-body") are not supported. Also the output HTML is not so neat. There is also no documentation of this yet; putting this out to see if @xi might find it useful :-).