feat: Build combined fixtures and expectation files (#4)
The idea behind this is to allow the tests not only to be run in a browser, but orchestrated from within the browser.

Currently the tests that rely on roles on the body ("application-alone-on-body-is-ignored" and "landmark-role-on-body") are not supported. Also the output HTML is not so neat, though can address this later.